# api/openapi.yaml
openapi: 3.1.0
info:
  title: Xminds Release Platform API
  version: 0.1.0-p0
  description: |
    Xminds Release Platform 管理面与工作负载 API 基础契约。
    管理人员使用当前启用 IAM OIDC 来源签发的 Bearer Token，自动化工作负载使用独立发行者与令牌。
  license:
    name: Apache-2.0
    identifier: Apache-2.0
servers:
  - url: /
    description: 当前部署来源
security:
  - oidcBearer: []
...
paths:
...
  /api/v1/audit-events:
    get:
      operationId: queryAuditEvents
      summary: 查询不可变审计事件
      description: 仅允许具有目标产品 audit.read 权限的审计员或管理员访问。
      tags: [Audit]
      x-required-action: audit.read
      parameters:
        - $ref: '#/components/parameters/ProductID'
        - name: actor_subject
          in: query
          schema:
            type: string
            maxLength: 512
        - name: action
          in: query
          schema:
            type: string
            maxLength: 128
        - name: outcome
          in: query
          schema:
            $ref: '#/components/schemas/AuditOutcome'
        - name: since
          in: query
          schema:
            type: string
            format: date-time
        - name: until
          in: query
          schema:
            type: string
            format: date-time
        - name: before_time
          in: query
          description: 与 before_id 同时提供的稳定时间游标。
          schema:
            type: string
            format: date-time
        - name: before_id
          in: query
          description: 与 before_time 同时提供的 UUID 游标。
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 100
      responses:
        '200':
          description: 审计事件页
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditEventPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/audit-exports:
    post:
      operationId: startAuditExport
      summary: 发起审计导出
      description: 持久化不可变导出请求，并在同一事务中创建 audit.export.v1 作业。
      tags: [Audit]
      x-required-action: audit.read
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuditExportRequest'
      responses:
        '202':
          description: 导出请求已受理
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditExport'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/audit-exports/{export_id}:
    get:
      operationId: getAuditExport
      summary: 获取审计导出状态
      tags: [Audit]
      x-required-action: audit.read
      parameters:
        - name: export_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 审计导出状态
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditExport'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: 导出请求不存在
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
...
components:
  parameters:
...
    ProductID:
      name: product_id
      in: query
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 128
...
  securitySchemes:
    oidcBearer:
      type: http
      scheme: bearer
      bearerFormat: OIDC-JWT
      description: 面向控制台用户的企业 OIDC 身份令牌；Issuer 来自当前 iam_login_state 精确指向的 enabled OIDC 来源，token_use 必须精确为 human。
    workloadBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 面向 CI/CD、客户端与同步执行器的短期工作负载令牌；使用独立静态受信发行者，token_use 必须精确为 workload，不与人员验签互相回退。
  responses:
...
    Unauthorized:
      description: Bearer Token 缺失、无效或已过期
      headers:
        WWW-Authenticate:
          schema:
            type: string
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Forbidden:
      description: 身份已验证但缺少产品范围或 audit.read 权限
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
...
  schemas:
...
    AuditOutcome:
      type: string
      enum: [success, denied, failed]
    AuditEvent:
      type: object
      additionalProperties: false
      required:
        - id
        - occurred_at
        - product_id
        - actor_subject
        - actor_kind
        - actor_provider
        - actor_roles
        - actor_product_ids
        - token_id
        - action
        - resource_type
        - resource_id
        - outcome
        - request_id
        - metadata
        - previous_hash
        - event_hash
      properties:
        id:
          type: string
          format: uuid
        occurred_at:
          type: string
          format: date-time
        product_id:
          type: string
        actor_subject:
          type: string
        actor_kind:
          type: string
          enum: [human, workload, local]
        actor_provider:
          type: string
          enum: ['', github-actions, github-enterprise-actions, gitlab-ci, api-token]
        actor_roles:
          type: array
          items:
            type: string
            enum: [admin, publisher, approver, auditor, viewer]
        actor_product_ids:
          type: array
          items:
            type: string
        token_id:
          type: string
        action:
          type: string
        resource_type:
          type: string
        resource_id:
          type: string
        outcome:
          $ref: '#/components/schemas/AuditOutcome'
        request_id:
          type: string
          format: uuid
        source_ip:
          type: string
        metadata:
          type: object
          additionalProperties: true
        previous_hash:
          type: string
          pattern: '^[0-9a-f]{64}$'
        event_hash:
          type: string
          pattern: '^[0-9a-f]{64}$'
    AuditEventPage:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AuditEvent'
        next_before_time:
          type: string
          format: date-time
        next_before_id:
          type: string
          format: uuid
    AuditExportRequest:
      type: object
      additionalProperties: false
      required: [product_id]
      properties:
        product_id:
          type: string
          minLength: 1
          maxLength: 128
        action:
          type: string
          maxLength: 128
        outcome:
          $ref: '#/components/schemas/AuditOutcome'
        since:
          type: string
          format: date-time
        until:
          type: string
          format: date-time
    AuditExport:
      type: object
      additionalProperties: false
      required: [id, product_id, requested_by, request_id, status, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        product_id:
          type: string
        requested_by:
          type: string
        request_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, completed, failed]
        object_key:
          type: string
        error_code:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    ProblemDetails:
      type: object
      additionalProperties: false
      required: [type, title, status, code, request_id]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        code:
          type: string
          pattern: '^[A-Z][A-Z0-9_]{2,127}$'
        request_id:
          type: string
          format: uuid